requirements.txt
mysql-connector-python
requests
mcp
